requirements.txt
environ-config==21.2.0
influxdb-client[ciso]==1.23.0
PyYAML==6.0
